requirements.txt
apt-clone
apturl
asgiref
beautifulsoup4
blinker
Brlapi
certifi
chardet
click
colorama
command-not-found
configobj
cryptography
cupshelpers
dbus-python
defer
distro
Django
django-environ
djangorestframework
environ
eyeD3
filelock
grpcio
gyp
httplib2
idna
ifaddr
IMDbPY
importlib-metadata
jeepney
keyring
launchpadlib
lazr.restfulclient
lazr.uri
louis
macaroonbakery
Mako
MarkupSafe
more-itertools
nemo-emblems
netaddr
netifaces
oauthlib
onboard
packaging
PAM
pexpect
Pillow
protobuf
psutil
ptyprocess
pycairo
pycups
pycurl
pyelftools
PyGObject
PyICU
pyinotify
PyJWT
pymacaroons
PyNaCl
pyparsing
pyparted
pyRFC3339
python-apt
python-debian
python-gnupg
python-magic
python-xlib
pytz
pyxdg
PyYAML
reportlab
requests
requests-file
SecretStorage
setproctitle
six
soupsieve
sqlparse
systemd-python
tinycss2
tldextract
typing_extensions
ubuntu-drivers-common
ufw
Unidecode
urllib3
wadllib
webencodings
xdg
xkit
xlrd
youtube-dl
zipp